feat(base_llm/): add transform request/response abstract methods to base config class
