--- litellm/llms/llm_base/transformation.py
@@ -1,43 +1,0 @@
-"""
-Common base config for all LLM providers
-"""
-
-import types
-from abc import ABC, abstractmethod
-
-
-class BaseConfig(ABC):
-    def __init__(self):
-        pass
-
-    @classmethod
-    def get_config(cls):
-        return {
-            k: v
-            for k, v in cls.__dict__.items()
-            if not k.startswith("__")
-            and not isinstance(
-                v,
-                (
-                    types.FunctionType,
-                    types.BuiltinFunctionType,
-                    classmethod,
-                    staticmethod,
-                ),
-            )
-            and v is not None
-        }
-
-    @abstractmethod
-    def get_supported_openai_params(self, model: str) -> list:
-        pass
-
-    @abstractmethod
-    def map_openai_params(
-        self,
-        non_default_params: dict,
-        optional_params: dict,
-        model: str,
-        drop_params: bool,
-    ) -> dict:
-        pass
